requirements.txt
celery==5.4.0
requests==2.32.3
python-dotenv==1.0.1
